website/requirements.txt
absl-py==0.11.0
amqp==5.0.2
asgiref==3.2.3
astunparse==1.6.3
async-generator==1.10
attrs==20.3.0
billiard==3.6.3.0
bleach==3.3.0
cachetools==4.1.1
celery==5.0.2
certifi==2020.11.8
cffi==1.14.5
chardet==3.0.4
click==7.1.2
click-didyoumean==0.0.3
click-repl==0.1.6
colorama==0.4.4
cryptography==3.4.6
decorator==4.4.2
defusedxml==0.6.0
Django==3.0
django-audiofield==0.10.0
django-celery==3.3.1
django-cors-headers==3.7.0
django-filter==2.4.0
django-paypal==1.1.1
django-rest-knox==4.1.0
django-rest-passwordreset==1.1.0
djangorestframework==3.12.2
entrypoints==0.3
gast==0.3.3
google-auth==1.4.2
google-auth-oauthlib==0.4.2
google-colab==1.0.0
google-pasta==0.2.0
grpcio==1.33.2
gunicorn==20.1.0
h5py==2.10.0
idna==2.8
importlib-metadata==3.1.0
ipykernel==4.6.1
ipython==5.5.0
ipython-genutils==0.2.0
Jinja2==2.11.3
joblib==0.17.0
jsonschema==3.2.0
jupyter-client==6.1.11
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
Keras==2.4.3
Keras-Preprocessing==1.1.2
kombu==5.0.2
Markdown==3.3.3
MarkupSafe==1.1.1
midi2audio==0.1.1
mistune==0.8.4
more-itertools==8.6.0
music21==6.3.0
nbclient==0.5.2
nbconvert==6.0.7
nbformat==5.1.2
nest-asyncio==1.5.1
notebook==5.2.2
numpy==1.18.0
oauthlib==3.1.0
opt-einsum==3.3.0
packaging==20.9
pandas==0.24.2
pandocfilters==1.4.3
pickleshare==0.7.5
Pillow==7.0.0
portpicker==1.2.0
prompt-toolkit==1.0.18
protobuf==3.14.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pydotplus==2.0.2
pydub==0.24.1
pygame==2.0.0
Pygments==2.8.0
pyparsing==2.4.5
pyrsistent==0.17.3
python-dateutil==2.8.1
pytz==2019.3
pywin32==300
PyYAML==5.3.1
pyzmq==22.0.3
requests==2.21.0
requests-oauthlib==1.3.0
rsa==4.6
scikit-learn==0.23.2
scipy==1.5.4
simplegeneric==0.8.1
six==1.12.0
sklearn==0.0
sqlparse==0.3.0
tensorboard==2.4.0
tensorboard-plugin-wit==1.7.0
tensorflow==2.3.1
tensorflow-estimator==2.3.0
termcolor==1.1.0
testpath==0.4.4
threadpoolctl==2.1.0
tornado==4.5.3
traitlets==5.0.5
urllib3==1.24.3
vine==5.0.0
virtualenv==16.7.9
wcwidth==0.2.5
webcolors==1.11.1
webencodings==0.5.1
Werkzeug==1.0.1
whitenoise==5.3.0
wrapt==1.12.1
zipp==3.4.0
